tests/wpt/tests/css/css-grid/abspos/grid-positioned-items-unknown-named-grid-line-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Grid positioned items unknown named grid line</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#abspos" title="9. Absolute Positioning">
<meta name="assert" content="This test checks that grid placement properties of absolutely positioned items can reference implicit grid lines.">
<link rel="stylesheet" href="/fonts/ahem.css">
<link rel="stylesheet" href="/css/support/grid.css">
<style>

.grid {
  grid-template-columns: 100px 200px;
  grid-template-rows: 50px 150px;
  width: 500px;
  height: 300px;
  border: 5px solid black;
  margin: 30px;
  padding: 15px;
  /* Ensures that the grid container is the containing block of the absolutely positioned grid children. */
  position: relative;
}

.absolute {
  position: absolute;
}

.startUnknownLine {
  background-color: blue;
  grid-column: foo / 3;
  grid-row: bar / 3;
}

.endUnknownLine {
  background-color: orange;
  grid-column: 1 / foo;
  grid-row: 1 / bar;
}

.startAndEndUnknownLines {
  background-color: green;
  grid-column: foo / bar;
  grid-row: foo / bar;
}

</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script type="text/javascript">
  setup({ explicit_done: true });
</script>

<body onload="document.fonts.ready.then(() => { checkLayout('.grid'); })">

<div id="log"></div>

<div class="grid">
  <div class="absolute sizedToGridArea startAndEndUnknownLines"
    data-offset-x="0" data-offset-y="0" data-expected-width="530" data-expected-height="330">
  </div>
  <div class="absolute sizedToGridArea endUnknownLine"
    data-offset-x="15" data-offset-y="15" data-expected-width="515" data-expected-height="315">
  </div>
  <div class="absolute sizedToGridArea startUnknownLine"
    data-offset-x="315" data-offset-y="215" data-expected-width="215" data-expected-height="115">
  </div>
</div>

<div class="grid directionRTL">
  <div class="absolute sizedToGridArea startAndEndUnknownLines"
    data-offset-x="0" data-offset-y="0" data-expected-width="530" data-expected-height="330">
  </div>
  <div class="absolute sizedToGridArea endUnknownLine"
    data-offset-x="0" data-offset-y="15" data-expected-width="515" data-expected-height="315">
  </div>
  <div class="absolute sizedToGridArea startUnknownLine"
    data-offset-x="0" data-offset-y="215" data-expected-width="215" data-expected-height="115">
  </div>
</div>

</body>
